% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchit_link.R
\name{cauchit_link}
\alias{cauchit_link}
\title{The Cauchit Link Function}
\usage{
cauchit_link()
}
\value{
An object of class \code{"link"} (see \code{\link{link-class}}) containing
the transformation functions and their exact analytical first and second derivatives.
}
\description{
Creates an object of class \code{"link"} implementing the Cauchit transformation.
This link function is based on the quantile function of the standard Cauchy distribution.
}
\details{
The Cauchit link is defined as \eqn{\eta = \tan(\pi(\theta - 0.5))}, corresponding to
\code{qcauchy(theta)}.

Unlike the Logit or Probit links, the Cauchit link has heavier tails. This makes it
particularly useful for modeling binary data where the probability approaches 0 or 1
more slowly, or when the data contains outliers that might influence the fit of light-tailed links.

The domain of \eqn{\theta} is \code{(0, 1)}.
}
\seealso{
\code{\link{link-class}}, \code{\link{logit_link}}, \code{\link[stats]{qcauchy}}
}
