% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglog_link.R
\name{cloglog_link}
\alias{cloglog_link}
\title{The Complementary Log-Log (ClogLog) Link Function}
\usage{
cloglog_link()
}
\value{
An object of class \code{"link"} (see \code{\link{link-class}}) containing
the transformation functions and their exact analytical first and second derivatives.
}
\description{
Creates an object of class \code{"link"} implementing the Complementary Log-Log transformation.
This link is asymmetric and is often used for modeling binary data where the probability of the event
approaches 1 slowly but approaches 0 sharply (or vice versa depending on formulation).
}
\details{
The ClogLog link is defined as \eqn{\eta = \log(-\log(1 - \theta))}.
The inverse link is \eqn{\theta = 1 - \exp(-\exp(\eta))}.

Unlike the symmetric Logit and Probit links, ClogLog is asymmetric. It is closely related to the
Gumbel distribution and is frequently used in survival analysis (proportional hazards models)
and for modeling rare events.

The domain of \eqn{\theta} is \code{(0, 1)}.
}
\seealso{
\code{\link{link-class}}, \code{\link{logit_link}}
}
