% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_link.r
\name{identity_link}
\alias{identity_link}
\title{The Identity Link Function}
\usage{
identity_link()
}
\value{
An object of class \code{"link"} (see \code{\link{link-class}}) containing
the transformation functions and their exact analytical first and second derivatives.
}
\description{
Creates an object of class \code{"link"} implementing the Identity transformation.
This is the canonical link function for mean parameter of the Normal (Gaussian) distribution.
}
\details{
The Identity link is defined simply as \eqn{\eta = \theta}.
Consequently, the inverse link is also \eqn{\theta = \eta}.

All first derivatives are constant (equal to 1), and all second derivatives are zero.

The domain of \eqn{\theta} is unbounded: \code{(-Inf, Inf)}.
}
\seealso{
\code{\link{link-class}}, \code{\link{logit_link}}
}
