% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_sq_link.R
\name{inverse_sq_link}
\alias{inverse_sq_link}
\title{The Inverse Square Link Function}
\usage{
inverse_sq_link()
}
\value{
An object of class \code{"link"} (see \code{\link{link-class}}) containing
the transformation functions and their exact analytical first and second derivatives.
}
\description{
Creates an object of class \code{"link"} implementing the Inverse Square transformation.
}
\details{
The Inverse Square link is defined as \eqn{\eta = 1 / \theta^2}.
The inverse link function is \eqn{\theta = 1 / \sqrt{\eta}}.
}
\seealso{
\code{\link{link-class}}, \code{\link{inverse_link}}, \code{\link[stats]{family}}
}
