% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_link.R
\name{log_link}
\alias{log_link}
\title{The Logarithmic Link Function}
\usage{
log_link()
}
\value{
An object of class \code{"link"} (see \code{\link{link-class}}) containing
the transformation functions and their exact analytical first and second derivatives.
}
\description{
Creates an object of class \code{"link"} implementing the natural logarithm transformation.
This is the canonical link function for mean of the Poisson distribution and is widely used for modeling count data
or non-negative continuous data with multiplicative effects.
}
\details{
The Log link is defined as \eqn{\eta = \log(\theta)}.
The inverse link is the exponential function \eqn{\theta = \exp(\eta)}.

A mathematical property of this link is that the inverse function is its own derivative.
Therefore, \eqn{\theta}, \eqn{d\theta/d\eta}, and \eqn{d^2\theta/d\eta^2} are all equal to \eqn{\exp(\eta)}.

The domain of \eqn{\theta} is \code{(0, Inf)}.
}
\seealso{
\code{\link{link-class}}, \code{\link{inverse_link}}
}
