% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_bounded_link.R
\name{lower_bounded_link}
\alias{lower_bounded_link}
\title{The Lower Bounded Link Function}
\usage{
lower_bounded_link(lwr = 0)
}
\arguments{
\item{lwr}{A numeric value specifying the lower bound. Defaults to 0 (which is equivalent to the \code{\link{log_link}}).}
}
\value{
An object of class \code{"link"} (see \code{\link{link-class}}).
}
\description{
Creates an object of class \code{"link"} implementing a logarithmic transformation shifted by a lower bound.
This allows modeling parameters that must be strictly greater than a specific value \code{lwr} (e.g., degrees of freedom > 2).
}
\details{
The Lower Bounded link is defined as \eqn{\eta = \log(\theta - \text{lwr})}.
The inverse link is \eqn{\theta = \exp(\eta) + \text{lwr}}.

The domain of \eqn{\theta} is \code{(lwr, Inf)}.
}
\seealso{
\code{\link{link-class}}, \code{\link{log_link}}, \code{\link{upper_bounded_link}}
}
