% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_methods.R
\name{print.link}
\alias{print.link}
\title{Print Method for Link Objects}
\usage{
\method{print}{link}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"link"}.}

\item{...}{Additional arguments passed to methods (currently unused).}
}
\value{
The function returns \code{x} invisibly.
}
\description{
A standard S3 print method for objects of class \code{"link"}.
It displays the name of the link function and the valid domain of the parameter \eqn{\theta}.
}
