% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_bounded_link.R
\name{upper_bounded_link}
\alias{upper_bounded_link}
\title{The Upper Bounded Link Function}
\usage{
upper_bounded_link(upr = 0)
}
\arguments{
\item{upr}{A numeric value specifying the upper bound.}
}
\value{
An object of class \code{"link"} (see \code{\link{link-class}}).
}
\description{
Creates an object of class \code{"link"} implementing a logarithmic transformation for parameters bounded from above.
This is useful for modeling parameters that must remain strictly below a certain threshold \code{upr}.
}
\details{
The Upper Bounded link is defined as \eqn{\eta = \log(\text{upr} - \theta)}.
The inverse link is \eqn{\theta = \text{upr} - \exp(\eta)}.

Note that as \eqn{\eta \to \infty}, \eqn{\theta \to -\infty}, and as \eqn{\eta \to -\infty}, \eqn{\theta \to \text{upr}}.

The domain of \eqn{\theta} is \code{(-Inf, upr)}.
}
\seealso{
\code{\link{link-class}}, \code{\link{lower_bounded_link}}
}
